refactor(header): move header files to specific folder
